pyproject.toml
[project]
name = "agentcore-registry-test"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aws-requests-auth>=0.4.3",
    "bedrock-agentcore>=1.6.0",
    "boto3>=1.42",
    "botocore[crt]>=1.42",
    "requests>=2.33.1",
    "strands-agents>=1.35.0",
]
